Moved MainWindow socket code to a separate class and added an exception for null socket and missing connection.  To do: adding send request code + potentially initializing socket with data?
