+++ ClientUI/MainWindow.xaml.cs
@@ -24,52 +24,18 @@ namespace ClientUI
     /// </summary>
     public partial class MainWindow : Window
     {
-        private Socket clientSocket;
+        // Potentially initialize? Will look later
+        // Create instance of SocketClient class
+        private SocketClient _socketClient;
         public MainWindow()
         {
             InitializeComponent();
         }
-        private void ConnectToServer()
+        // On program window closing, close socket connection.
+        private void Window_Closing(object sender, System.ComponentModel.CancelEventArgs e)
         {
-            try
-            {
-                var host = Dns.GetHostEntry("localhost");
-                var ipAddress = host.AddressList[0];
-                var remoteEndPoint = new IPEndPoint(ipAddress, 11000);
+            _socketClient.Close();
+        }
 
-                clientSocket = new Socket(ipAddress.AddressFamily, SocketType.Stream, ProtocolType.Tcp);
-                clientSocket.Connect(remoteEndPoint);
-                
-            }
-            catch (Exception ex)
-            {
-               MessageBox.Show(ex.Message);
-            }
-        }
-        private void SendCurrentUsage(object sender, RoutedEventArgs e)
-        {
-            
-        }
-        private void SendRequest(RequestType requestType,CurrentUsageRequest currentUsageRequest)
-        {
-            if(clientSocket == null || !clientSocket.Connected)
-            {
-                //Display error message
-                return;
-            }
-            try
-            {
-                var request = new ServerRequest
-                {
-                    Type = requestType,
-                    Data = JsonConvert.SerializeObject(currentUsageRequest)
-                };
-
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show(ex.Message);
-            }
-        }
     }
 }
